adds readme and stuff

+++ install.sh
@@ -21,6 +21,13 @@ fi
 
 # Copy all of the xctemplate folders into the install directory.
 #cp -r *.xctemplate "$installDirectory"
-echo "Templates will be linked from $current_dir into ${template_dir}${template_name}"
-ln -s "$current_dir" "${template_dir}${template_name}"
+if [ ! -L "${template_dir}${template_name}" ]; then
 
+  echo "Templates will be linked from $current_dir into ${template_dir}${template_name}"
+  ln -s "$current_dir" "${template_dir}${template_name}"
+
+else
+
+  echo "Symbolic link $template_name already exists in the $template_dir"
+
+fi
